Added more files into a list of files to be ignored while parsing the repo
