+++ backend/config/ignore_file_list.json
@@ -1,0 +1,126 @@
+{
+    "files": [
+      ".env",
+      ".env.local",
+      ".env.development",
+      ".env.production",
+      ".env.test",
+      ".envrc",
+      ".npmrc",
+      ".yarnrc",
+      ".yarnrc.yml",
+      ".nvmrc",
+      ".watchmanconfig",
+      ".gitignore",
+      ".gitattributes",
+      "Dockerfile",
+      "docker-compose.yml",
+      ".dockerignore",
+      ".travis.yml",
+      ".appveyor.yml",
+      ".gitlab-ci.yml",
+      ".codeclimate.yml",
+      ".coveralls.yml",
+      "Makefile",
+      "CMakeLists.txt",
+      "package.json",
+      "package-lock.json",
+      "yarn.lock",
+      "pnpm-lock.yaml",
+      "lerna.json",
+      ".babelrc",
+      ".babelrc.json",
+      "babel.config.js",
+      ".eslintrc",
+      ".eslintrc.json",
+      ".eslintrc.js",
+      ".eslintrc.yml",
+      ".eslintrc.yaml",
+      ".prettierrc",
+      ".prettierrc.json",
+      ".prettierrc.yml",
+      ".prettierrc.yaml",
+      ".prettierrc.js",
+      ".prettierignore",
+      ".stylelintrc",
+      ".stylelintrc.json",
+      ".stylelintrc.yml",
+      ".stylelintrc.yaml",
+      ".stylelintrc.js",
+      ".jshintrc",
+      "tsconfig.json",
+      "jsconfig.json",
+      "requirements.txt",
+      "pyproject.toml",
+      "setup.cfg",
+      "tox.ini",
+      ".flake8",
+      ".pylintrc",
+      ".condarc",
+      "Gemfile",
+      "Gemfile.lock",
+      ".rubocop.yml",
+      "Rakefile",
+      "composer.json",
+      "composer.lock",
+      "phpunit.xml",
+      "phpcs.xml",
+      "pom.xml",
+      "build.gradle",
+      "settings.gradle",
+      "gradle.properties",
+      "build.xml",
+      "*.iml",
+      ".project",
+      ".classpath",
+      "configure.ac",
+      ".clang-format",
+      ".clang-tidy",
+      "Cargo.toml",
+      "Cargo.lock",
+      "go.mod",
+      "go.sum",
+      "Gopkg.toml",
+      "Gopkg.lock",
+      "pubspec.yaml",
+      "pubspec.lock",
+      "Podfile",
+      "Podfile.lock",
+      "project.pbxproj",
+      "build.sbt",
+      "project/build.properties",
+      "project/plugins.sbt",
+      "config.ini",
+      "config.json",
+      "config.js",
+      "config.xml",
+      "app.config",
+      "web.config",
+      "settings.json",
+      "local.settings.json",
+      ".bazelrc",
+      ".buckconfig"
+    ],
+    "directories": [
+      ".git",
+      ".circleci",
+      ".github/workflows",
+      ".vscode",
+      ".idea",
+      "config",
+      "configs",
+      "settings",
+      ".husky",
+      ".vagrant",
+      ".mvn",
+      "node_modules",
+      "bower_components",
+      "vendor",
+      "dist",
+      "build",
+      "target",
+      "out",
+      "public"
+    ]
+  }
+
